+++ src/main/java/com/easypost/model/Order.java
@@ -479,4 +479,37 @@ public final class Order extends EasyPostResource {
         this.merge(this, response);
         return this;
     }
+
+    /**
+     * Get the lowest rate for this Order.
+     *
+     * @return lowest Rate object
+     * @throws EasyPostException when the request fails.
+     */
+    public Rate lowestRate() throws EasyPostException {
+        return this.lowestRate(null, null);
+    }
+
+    /**
+     * Get the lowest rate for this order.
+     *
+     * @param carriers the carriers to use in the query.
+     * @return Rate object
+     * @throws EasyPostException when the request fails.
+     */
+    public Rate lowestRate(final List<String> carriers) throws EasyPostException {
+        return this.lowestRate(carriers, null);
+    }
+
+    /**
+     * Get the lowest rate for this Order.
+     *
+     * @param carriers the carriers to use in the filter.
+     * @param services the services to use in the filter.
+     * @return lowest Rate object
+     * @throws EasyPostException when the request fails.
+     */
+    public Rate lowestRate(final List<String> carriers, final List<String> services) throws EasyPostException {
+        return Utilities.getLowestObjectRate(this.rates, carriers, services);
+    }
 }