+++ src/test/java/com/easypost/OrderTest.java
@@ -1,17 +1,22 @@
 package com.easypost;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.List;
-
 import com.easypost.exception.EasyPostException;
 import com.easypost.model.Order;
-import com.easypost.model.Shipment;
 import com.easypost.model.Rate;
-
-import org.junit.jupiter.api.Test;
+import com.easypost.model.Shipment;
 import org.junit.jupiter.api.BeforeAll;
-import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class OrderTest {
     private static Order globalOrder;
@@ -91,5 +96,34 @@ public class OrderTest {
         for (Shipment shipment : shipments) {
             assertNotNull(shipment.getPostageLabel());
         }
+    }
+
+    /**
+     * Test getting the lowest rate of an Order.
+     *
+     * @throws EasyPostException when the request fails.
+     */
+    @Test
+    public void testLowestRate() throws EasyPostException {
+        Order order = Order.create(Fixture.basicOrder());
+
+        // Test lowest rate with no filters
+        Rate lowestRate = order.lowestRate();
+        assertEquals("First", lowestRate.getService());
+        assertEquals(5.49, lowestRate.getRate(), 0.01);
+        assertEquals("USPS", lowestRate.getCarrier());
+
+        // Test lowest rate with service filter (this rate is higher than the lowest but should filter)
+        List<String> services = new ArrayList<>(Arrays.asList("Priority"));
+        Rate lowestRateService = order.lowestRate(null, services);
+        assertEquals("Priority", lowestRateService.getService());
+        assertEquals(7.37, lowestRateService.getRate(), 0.01);
+        assertEquals("USPS", lowestRateService.getCarrier());
+
+        // Test lowest rate with carrier filter (should error due to bad carrier)
+        List<String> carriers = new ArrayList<>(Arrays.asList("BAD CARRIER"));
+        assertThrows(EasyPostException.class, () -> {
+            order.lowestRate(carriers, null);
+        });
     }
 }